translation/requirements_voice.txt
python-docx
gTTS
pydub
openai
google-generativeai
python-dotenv
openai-whisper
